requirements.txt
# schema_design
psycopg2-binary==2.9.3
faker==13.15


# movies_admin
django==3.2
django-split-settings


# sqlite_to_postgres


# Workflow
flake8==4.0
wemake-python-styleguide==0.16.1
flake8-bandit==3.0.0
bandit==1.7.4
pre-commit==2.20
python-dotenv==0.20.0